+++ tket/test/src/Circuit/test_Boxes.cpp
@@ -21,8 +21,12 @@
 #include "tket/Circuit/Boxes.hpp"
 #include "tket/Circuit/CircUtils.hpp"
 #include "tket/Circuit/Circuit.hpp"
+#include "tket/Circuit/DiagonalBox.hpp"
+#include "tket/Circuit/Multiplexor.hpp"
 #include "tket/Circuit/PauliExpBoxes.hpp"
 #include "tket/Circuit/Simulation/CircuitSimulator.hpp"
+#include "tket/Circuit/StatePreparation.hpp"
+#include "tket/Circuit/ToffoliBox.hpp"
 #include "tket/Converters/PhasePoly.hpp"
 #include "tket/Gate/SymTable.hpp"
 
@@ -833,8 +837,16 @@ SCENARIO("Checking equality", "[boxes]") {
     u.add_op<unsigned>(OpType::CX, {0, 1});
     const CircBox ubox(u);
 
+    Circuit u2(2);
+    u2.add_op<unsigned>(OpType::Ry, -0.35, {0});
+    u2.add_op<unsigned>(OpType::CX, {0, 1});
+    const CircBox ubox2(u2);
+
     WHEN("both arguments are equal") { REQUIRE(ubox == ubox); }
-    WHEN("both arguments are different") { REQUIRE(ubox != CircBox(u)); }
+    WHEN("different ids but equivalent inner circuits") {
+      REQUIRE(ubox == CircBox(u));
+    }
+    WHEN("different inner circuits") { REQUIRE(ubox != ubox2); }
   }
   GIVEN("Unitary1qBox") {
     Circuit setup(1);
@@ -843,11 +855,16 @@ SCENARIO("Checking equality", "[boxes]") {
     Unitary1qBox mbox(m);
 
     WHEN("both arguments are equal") { REQUIRE(mbox == mbox); }
-    WHEN("both arguments are different") {
-      setup.add_op<unsigned>(OpType::TK1, {0.2374, 1.0353, 0.5372}, {0});
+    WHEN("different ids but matrices are equal") {
       Eigen::Matrix2cd m2 = tket_sim::get_unitary(setup);
       Unitary1qBox mbox2(m2);
-      REQUIRE(mbox != mbox2);
+      REQUIRE(mbox == mbox2);
+    }
+    WHEN("both arguments are different") {
+      setup.add_op<unsigned>(OpType::TK1, {0.2374, 1.0353, 0.5372}, {0});
+      Eigen::Matrix2cd m3 = tket_sim::get_unitary(setup);
+      Unitary1qBox mbox3(m3);
+      REQUIRE(mbox != mbox3);
     }
   }
   GIVEN("Unitary2qBox") {
@@ -858,11 +875,16 @@ SCENARIO("Checking equality", "[boxes]") {
     Unitary2qBox mbox(m);
 
     WHEN("both arguments are equal") { REQUIRE(mbox == mbox); }
-    WHEN("both arguments are different") {
-      setup.add_op<unsigned>(OpType::CX, {1, 0});
+    WHEN("different ids but matrices are equal") {
       Eigen::Matrix4cd m2 = tket_sim::get_unitary(setup);
       Unitary2qBox mbox2(m2);
-      REQUIRE(mbox != mbox2);
+      REQUIRE(mbox == mbox2);
+    }
+    WHEN("both arguments are different") {
+      setup.add_op<unsigned>(OpType::CX, {1, 0});
+      Eigen::Matrix4cd m3 = tket_sim::get_unitary(setup);
+      Unitary2qBox mbox3(m3);
+      REQUIRE(mbox != mbox3);
     }
   }
   GIVEN("Unitary3qBox") {
@@ -874,11 +896,16 @@ SCENARIO("Checking equality", "[boxes]") {
     Unitary3qBox mbox(m);
 
     WHEN("both arguments are equal") { REQUIRE(mbox == mbox); }
-    WHEN("both arguments are different") {
-      setup.add_op<unsigned>(OpType::CX, {0, 2});
+    WHEN("different ids but matrices are equal") {
       Eigen::MatrixXcd m2 = tket_sim::get_unitary(setup);
       Unitary3qBox mbox2(m2);
-      REQUIRE(mbox != mbox2);
+      REQUIRE(mbox == mbox2);
+    }
+    WHEN("both arguments are different") {
+      setup.add_op<unsigned>(OpType::CX, {0, 2});
+      Eigen::MatrixXcd m3 = tket_sim::get_unitary(setup);
+      Unitary3qBox mbox3(m3);
+      REQUIRE(mbox != mbox3);
     }
   }
   GIVEN("ExpBox") {
@@ -888,9 +915,13 @@ SCENARIO("Checking equality", "[boxes]") {
         4., 2. + 3. * i_, 5.;
     ExpBox ebox(A, -0.5);
     WHEN("both arguments are equal") { REQUIRE(ebox == ebox); }
+    WHEN("different ids but matrices are equal") {
+      ExpBox ebox2(A, -0.5);
+      REQUIRE(ebox == ebox2);
+    }
     WHEN("both arguments are different") {
-      ExpBox ebox2(A, -0.2);
-      REQUIRE(ebox != ebox2);
+      ExpBox ebox3(A, -0.2);
+      REQUIRE(ebox != ebox3);
     }
   }
   GIVEN("Pauli gadgets") {
@@ -903,13 +934,26 @@ SCENARIO("Checking equality", "[boxes]") {
     }
   }
   GIVEN("QControlBox") {
-    Op_ptr op = get_op_ptr(OpType::X);
+    Circuit u(2);
+    u.add_op<unsigned>(OpType::CX, {0, 1});
+    Op_ptr op = std::make_shared<CircBox>(CircBox(u));
     QControlBox qcbox(op);
     WHEN("both arguments are equal") { REQUIRE(qcbox == qcbox); }
-    WHEN("both arguments are different") {
-      Op_ptr op2 = get_op_ptr(OpType::Y);
+    WHEN("different ids but equivalent ops") {
+      Circuit u2(2);
+      u2.add_op<unsigned>(OpType::CX, {0, 1});
+      Op_ptr op2 = std::make_shared<CircBox>(CircBox(u2));
       QControlBox qcbox2(op2);
-      REQUIRE(qcbox != qcbox2);
+      REQUIRE(qcbox == qcbox2);
+    }
+    WHEN("different ids, equivalent ops, but different types") {
+      Op_ptr op3 = get_op_ptr(OpType::CX);
+      REQUIRE(qcbox != QControlBox(op3));
+    }
+    WHEN("both arguments are different") {
+      Op_ptr op4 = get_op_ptr(OpType::Y);
+      QControlBox qcbox4(op4);
+      REQUIRE(qcbox != qcbox4);
     }
   }
   GIVEN("PhasePolyBox") {
@@ -954,6 +998,208 @@ SCENARIO("Checking equality", "[boxes]") {
     CHECK(g1 != g1_wrong);
     CHECK(g1_repeated != g1_wrong);
     CHECK_THROWS_AS(CustomGate(nullptr, {param3}), std::runtime_error);
+  }
+  GIVEN("ProjectorAssertionBox") {
+    Eigen::MatrixXcd bell(4, 4);
+    bell << 0.5, 0, 0, 0.5, 0, 0, 0, 0, 0, 0, 0, 0, 0.5, 0, 0, 0.5;
+    ProjectorAssertionBox box(bell);
+    WHEN("both arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but equivalent projectors") {
+      REQUIRE(box == ProjectorAssertionBox(bell));
+    }
+    WHEN("different projectors") {
+      Eigen::MatrixXcd p(4, 4);
+      p << 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0;
+      REQUIRE(box != ProjectorAssertionBox(p));
+    }
+  }
+  GIVEN("StabiliserAssertionBox") {
+    PauliStabiliser p1 = {{Pauli::X, Pauli::X}, true};
+    PauliStabiliser p2 = {{Pauli::Z, Pauli::Z}, true};
+    PauliStabiliser p3 = {{Pauli::Z, Pauli::Z}, false};
+    StabiliserAssertionBox box({p1, p2});
+    WHEN("both arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but equivalent stabilisers") {
+      REQUIRE(box == StabiliserAssertionBox({p1, p2}));
+    }
+    WHEN("different stabilisers") {
+      REQUIRE(box != StabiliserAssertionBox({p1, p3}));
+    }
+  }
+  GIVEN("DiagonalBox") {
+    Eigen::Vector2cd diag(i_, 1);
+    DiagonalBox box(diag);
+    WHEN("all arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but other args are equal") {
+      DiagonalBox box2(diag);
+      REQUIRE(box == box2);
+    }
+    WHEN("arguments are different") {
+      DiagonalBox box3(diag, false);
+      REQUIRE(box != box3);
+    }
+  }
+  GIVEN("MultiplexorBox") {
+    ctrl_op_map_t op_map = {{{1}, get_op_ptr(OpType::H)}};
+    MultiplexorBox box(op_map);
+    WHEN("all arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but other args are equal") {
+      ctrl_op_map_t op_map2 = {{{1}, get_op_ptr(OpType::H)}};
+      MultiplexorBox box2(op_map2);
+      REQUIRE(box == box2);
+    }
+    WHEN("arguments are different") {
+      ctrl_op_map_t op_map2 = {{{0}, get_op_ptr(OpType::H)}};
+      MultiplexorBox box3(op_map2);
+      REQUIRE(box != box3);
+    }
+  }
+  GIVEN("MultiplexedRotationBox") {
+    ctrl_op_map_t op_map = {{{1}, get_op_ptr(OpType::Rz, 0.7)}};
+    MultiplexedRotationBox box(op_map);
+    WHEN("all arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but other args are equal") {
+      ctrl_op_map_t op_map2 = {{{1}, get_op_ptr(OpType::Rz, 0.7)}};
+      MultiplexedRotationBox box2(op_map2);
+      REQUIRE(box == box2);
+    }
+    WHEN("arguments are different") {
+      ctrl_op_map_t op_map2 = {{{0}, get_op_ptr(OpType::Rz, 0.7)}};
+      MultiplexedRotationBox box3(op_map2);
+      REQUIRE(box != box3);
+    }
+  }
+  GIVEN("MultiplexedU2Box") {
+    ctrl_op_map_t op_map = {{{1}, get_op_ptr(OpType::H)}};
+    MultiplexedU2Box box(op_map);
+    WHEN("all arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but other args are equal") {
+      ctrl_op_map_t op_map2 = {{{1}, get_op_ptr(OpType::H)}};
+      MultiplexedU2Box box2(op_map2);
+      REQUIRE(box == box2);
+    }
+    WHEN("arguments are different") {
+      MultiplexedU2Box box3(op_map, false);
+      REQUIRE(box != box3);
+    }
+  }
+  GIVEN("MultiplexedTensoredU2Box") {
+    ctrl_tensored_op_map_t op_map;
+    op_map.insert({{0, 0}, {get_op_ptr(OpType::X), get_op_ptr(OpType::X)}});
+    MultiplexedTensoredU2Box box(op_map);
+    WHEN("all arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but other args are equal") {
+      ctrl_tensored_op_map_t op_map2;
+      op_map2.insert({{0, 0}, {get_op_ptr(OpType::X), get_op_ptr(OpType::X)}});
+      MultiplexedTensoredU2Box box2(op_map2);
+      REQUIRE(box == box2);
+    }
+    WHEN("arguments are different") {
+      op_map.insert({{0, 1}, {get_op_ptr(OpType::H), get_op_ptr(OpType::X)}});
+      MultiplexedTensoredU2Box box3(op_map);
+      REQUIRE(box != box3);
+    }
+  }
+  GIVEN("StatePreparationBox") {
+    Eigen::Vector2cd diag(0, 1);
+    StatePreparationBox box(diag);
+    WHEN("all arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but other args are equal") {
+      StatePreparationBox box2(diag);
+      REQUIRE(box == box2);
+    }
+    WHEN("arguments are different") {
+      StatePreparationBox box3(diag, false, true);
+      REQUIRE(box != box3);
+    }
+  }
+  GIVEN("ToffoliBox") {
+    state_perm_t perm;
+    perm[{0, 1}] = {1, 0};
+    perm[{1, 0}] = {0, 1};
+    ToffoliBox box(perm);
+    WHEN("all arguments are equal") { REQUIRE(box == box); }
+    WHEN("different ids but other args are equal") {
+      ToffoliBox box2(perm);
+      REQUIRE(box == box2);
+    }
+    WHEN("permutations are equivalent") {
+      state_perm_t perm2;
+      perm2[{0, 1}] = {1, 0};
+      perm2[{1, 0}] = {0, 1};
+      perm2[{1, 1}] = {1, 1};
+      ToffoliBox box2(perm2);
+      REQUIRE(box == box2);
+    }
+    WHEN("permutations are not equivalent") {
+      state_perm_t perm2;
+      perm2[{0, 1}] = {1, 0};
+      perm2[{1, 0}] = {0, 1};
+      perm2[{1, 1}] = {0, 0};
+      perm2[{0, 0}] = {1, 1};
+      ToffoliBox box2(perm2);
+      REQUIRE(box != box2);
+    }
+    WHEN("permutations are not equivalent case 2") {
+      state_perm_t perm2;
+      perm2[{1, 1}] = {0, 0};
+      perm2[{0, 0}] = {1, 1};
+      ToffoliBox box2(perm2);
+      REQUIRE(box != box2);
+    }
+    WHEN("arguments are different") {
+      ToffoliBox box2(perm, ToffoliBoxSynthStrat::Cycle);
+      REQUIRE(box != box2);
+    }
+  }
+  GIVEN("PauliExpBox") {
+    PauliExpBox pbox({Pauli::X, Pauli::Y, Pauli::Z}, 0.8);
+    WHEN("both arguments are equal") { REQUIRE(pbox == pbox); }
+    WHEN("different ids but same arguments") {
+      REQUIRE(pbox == PauliExpBox({Pauli::X, Pauli::Y, Pauli::Z}, 0.8));
+    }
+    WHEN("different ids, equivalent angle") {
+      REQUIRE(pbox == PauliExpBox({Pauli::X, Pauli::Y, Pauli::Z}, 4.8));
+    }
+    WHEN("different arguments") {
+      REQUIRE(pbox != PauliExpBox({Pauli::X, Pauli::Y, Pauli::Z}, 0.9));
+    }
+  }
+  GIVEN("PauliExpPairBox") {
+    PauliExpPairBox pbox({Pauli::X}, 1.0, {Pauli::I}, 0.0);
+    WHEN("both arguments are equal") { REQUIRE(pbox == pbox); }
+    WHEN("different ids but same arguments") {
+      REQUIRE(pbox == PauliExpPairBox({Pauli::X}, 1.0, {Pauli::I}, 0.0));
+    }
+    WHEN("different ids, equivalent angle") {
+      REQUIRE(pbox == PauliExpPairBox({Pauli::X}, 1.0, {Pauli::I}, 4.0));
+    }
+    WHEN("different arguments") {
+      REQUIRE(pbox != PauliExpPairBox({Pauli::X}, -1.0, {Pauli::I}, 0.0));
+    }
+  }
+  GIVEN("PauliExpCommutingSetBox") {
+    PauliExpCommutingSetBox pbox(
+        {{{Pauli::X}, 1.0}, {{Pauli::I}, 1.2}, {{Pauli::I}, -0.5}});
+    WHEN("both arguments are equal") { REQUIRE(pbox == pbox); }
+    WHEN("different ids but same arguments") {
+      REQUIRE(
+          pbox ==
+          PauliExpCommutingSetBox(
+              {{{Pauli::X}, 1.0}, {{Pauli::I}, 1.2}, {{Pauli::I}, -0.5}}));
+    }
+    WHEN("different ids, equivalent angles") {
+      REQUIRE(
+          pbox ==
+          PauliExpCommutingSetBox(
+              {{{Pauli::X}, -3.0}, {{Pauli::I}, 5.2}, {{Pauli::I}, -0.5}}));
+    }
+    WHEN("different arguments") {
+      REQUIRE(
+          pbox !=
+          PauliExpCommutingSetBox(
+              {{{Pauli::Y}, 1.0}, {{Pauli::I}, 1.2}, {{Pauli::I}, -0.5}}));
+    }
   }
 }